Revert "ci(solana): run devnet-only examples on fork"
This reverts commit 5933ff2b87a3a23589396c1cb88d77cf38295f04.
