+++ examples/.env.example
@@ -68,12 +68,6 @@ RPC_URL=https://mainnet.base.org
 # SOLANA_INITIALIZER_PROGRAM_ID=INITIALIZER_PROGRAM_ID
 # SOLANA_CPMM_MIGRATOR_PROGRAM_ID=CPMM_MIGRATOR_PROGRAM_ID
 # SOLANA_DOPPLER_LAUNCH_HOOK_V1_PROGRAM_ID=DOPPLER_LAUNCH_HOOK_V1_PROGRAM_ID
-# SOLANA_DOPPLER_LAUNCH_HOOK_V2_PROGRAM_ID=DOPPLER_LAUNCH_HOOK_V2_PROGRAM_ID
-# SOLANA_DOPPLER_REHYPE_ROUTER_V1_PROGRAM_ID=DOPPLER_REHYPE_ROUTER_V1_PROGRAM_ID
-# SOLANA_VESTING_PROGRAM_ID=DOPPLER_VESTING_PROGRAM_ID
-# Optional overrides for examples/solana-vesting-launch.ts:
-# SOLANA_VESTING_CLIFF_SECONDS=604800
-# SOLANA_VESTING_DURATION_SECONDS=2592000
 # Required by examples/solana-vesting-claim.ts:
 # SOLANA_VESTING_LAUNCH=LAUNCH_ACCOUNT
 # SOLANA_VESTING_BASE_MINT=BASE_MINT
