+++ examples/solana-vesting-launch.ts
@@ -1,9 +1,7 @@
 /** Creates a non-migrating launch with an immutable token vesting allocation. */
 import './env.js';
 
-import { writeFileSync } from 'node:fs';
-
-import { address, generateKeyPairSigner } from '@solana/kit';
+import { generateKeyPairSigner } from '@solana/kit';
 
 import { vesting } from '../src/solana/index.js';
 import {
@@ -24,52 +22,14 @@ const VESTED_SUPPLY = (BASE_TOTAL_SUPPLY * 20n) / 100n;
 const DAY_SECONDS = 86_400n;
 const LAMPORTS_PER_SOL = 1_000_000_000n;
 
-function nonnegativeBigIntEnv(name: string, defaultValue: bigint): bigint {
-  const text = process.env[name]?.trim();
-  if (!text) {
-    return defaultValue;
-  }
-
-  let value: bigint;
-  try {
-    value = BigInt(text);
-  } catch {
-    throw new Error(`${name} must be an integer`);
-  }
-  if (value < 0n) {
-    throw new Error(`${name} must not be negative`);
-  }
-  return value;
-}
-
-function writeStateOutput(state: { launch: string; baseMint: string }): void {
-  const outputPath = process.env.SOLANA_VESTING_STATE_OUTPUT?.trim();
-  if (!outputPath) {
-    return;
-  }
-  writeFileSync(outputPath, `${JSON.stringify(state, null, 2)}\n`, 'utf8');
-}
-
 async function main(): Promise<void> {
   const payer = await loadKeypairSignerFromEnv();
   const { rpc, rpcSubscriptions, network } = createSolanaClientsFromEnv();
-  assertSolanaExampleNetwork(network, ['devnet', 'custom']);
+  assertSolanaExampleNetwork(network, ['custom']);
   const deployment = await getSolanaCpmmDeploymentFromEnv(network);
   const baseMint = await generateKeyPairSigner();
   const baseVault = await generateKeyPairSigner();
   const quoteVault = await generateKeyPairSigner();
-  const cliffSeconds = nonnegativeBigIntEnv(
-    'SOLANA_VESTING_CLIFF_SECONDS',
-    7n * DAY_SECONDS,
-  );
-  const durationSeconds = nonnegativeBigIntEnv(
-    'SOLANA_VESTING_DURATION_SECONDS',
-    30n * DAY_SECONDS,
-  );
-  const vestingProgramText = process.env.SOLANA_VESTING_PROGRAM_ID?.trim();
-  const vestingProgram = vestingProgramText
-    ? address(vestingProgramText)
-    : undefined;
 
   const prepared = await vesting.prepareLaunch({
     deployment,
@@ -92,7 +52,12 @@ async function main(): Promise<void> {
       swapFeeBps: DEFAULT_SWAP_FEE_BPS,
     },
     vesting: {
-      schedules: [{ cliffSeconds, durationSeconds }],
+      schedules: [
+        {
+          cliffSeconds: 7n * DAY_SECONDS,
+          durationSeconds: 30n * DAY_SECONDS,
+        },
+      ],
       allocations: [
         {
           beneficiary: payer.address,
@@ -103,7 +68,6 @@ async function main(): Promise<void> {
     },
     metadata: DEFAULT_TEST_METADATA,
     feeBeneficiaries: [{ wallet: payer.address, shareBps: 10_000 }],
-    vestingProgram,
   });
 
   const configSignature = await sendInstructions({
@@ -124,11 +88,6 @@ async function main(): Promise<void> {
     rpcSubscriptions,
     payer,
     instructions: [prepared.fundVestingInstruction],
-  });
-
-  writeStateOutput({
-    launch: prepared.launchAddresses.launch,
-    baseMint: baseMint.address,
   });
 
   console.log('Vesting launch created and funded:');
